requirements.txt
huawei-lte-api
nuitka
python-dotenv